requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==4.9.4
pandas==1.5.3
snowflake-connector-python==3.10.1
snowflake-connector-python[pandas]==3.10.1
dbt-core==1.7.17
dbt-snowflake==1.7.3
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.2.1
pytest==7.4.4
pytest-mock==3.12.0